@@ -1,10 +1,10 @@
+using FreedomCalculator2.Exceptions;
+using FreedomCalculator2.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Xml.Linq;
-using FreedomCalculator2.Exceptions;
-using FreedomCalculator2.Models;
 
 namespace FreedomCalculator2
 {
