@@ -5,7 +5,7 @@ namespace FreedomCalculator2.Models
 {
 	public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
     {
-		public ApplicationDbContext(DbContextOptions options) : base(options)
+		public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options)
 		{ }
 
 		protected override void OnModelCreating(ModelBuilder builder)
