@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
 namespace FreedomCalculator2.Models
@@ -12,6 +11,7 @@ namespace FreedomCalculator2.Models
         public virtual Budget Budget { get; set; }
         public int ExpenseId { get; set; }
         public virtual Expense Expense { get; set; }
+        [Column(TypeName = "decimal(18,2)")]
         public decimal Projected { get; set; }
         public List<BudgetExpenseItem> BudgetExpenseItems { get; set; }
     }
